requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
APScheduler==3.8.1
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.2.0
beautifulsoup4==4.9.3
bilibili-api==5.1.1
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.11.0
fastapi==0.63.0
feedparser==6.0.8
fonttools==4.28.2
gino==1.0.1
greenlet==1.0.0
h11==0.12.0
httpcore==0.12.3
httptools==0.1.2
httpx==0.17.1
idna==3.3
ImageHash==4.2.1
jieba==0.42.1
kiwisolver==1.3.2
loguru==0.5.3
lxml==4.6.4
matplotlib==3.5.0
multidict==5.2.0
nonebot-adapter-cqhttp==2.0.0a12
nonebot-plugin-apscheduler==0.1.2
nonebot2==2.0.0a13.post1
numpy==1.21.4
opencv-python==4.5.1.48
packaging==21.3
Pillow==8.2.0
playwright==1.10.0
psutil==5.8.0
pydantic==1.8.2
pyee==8.2.2
pygtrie==2.4.2
pyparsing==3.0.6
pypinyin==0.41.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==5.4.1
retrying==1.3.3
rfc3986==1.5.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.7.3
setuptools-scm==6.3.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
starlette==0.13.6
tomli==1.2.2
tomlkit==0.7.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
yarl==1.7.2
